chore: centralize requested token attachment
Move the artifact bundle type to models and keep sandbox token enrichment behind a shared host-side helper.
