+++ letta_evals/models/__init__.py
@@ -25,6 +25,7 @@ from letta_evals.models.results import (
     LettaMessageUnion,
     PerTurnGrade,
     RewardOutput,
+    RunArtifacts,
     SampleResult,
     TargetResult,
     Timing,
@@ -75,6 +76,7 @@ __all__ = [
     "PerTurnGrade",
     "GradeResult",
     "RewardOutput",
+    "RunArtifacts",
     "Usage",
     "Timing",
     "TimingStats",
