+++ letta_evals/runner.py
@@ -4,7 +4,7 @@ import os
 import time
 from collections import defaultdict
 from pathlib import Path
-from typing import Dict, List, Optional, TypedDict
+from typing import Dict, List, Optional
 
 import anyio
 import yaml
@@ -19,13 +19,12 @@ from letta_evals.graders.rubric import RubricGrader
 from letta_evals.graders.tool import ToolGrader
 from letta_evals.metrics import summarize_model, summarize_runs
 from letta_evals.models import (
-    AgentState,
     Error,
-    LettaMessageUnion,
     MetricRewardSpec,
     ModelJudgeGraderSpec,
     ModelRun,
     ModelSummary,
+    RunArtifacts,
     RunnerResult,
     Sample,
     SampleId,
@@ -34,7 +33,6 @@ from letta_evals.models import (
     Summary,
     Timing,
     ToolGraderSpec,
-    TurnTokenData,
     Usage,
 )
 from letta_evals.pricing import calculate_cost_from_agent_usage
@@ -55,17 +53,6 @@ logger = logging.getLogger(__name__)
 
 # Sentinel model identifier used when a suite has no model_handles.
 DEFAULT_MODEL_ID = "default"
-
-
-class RunArtifacts(TypedDict):
-    """Target execution metadata plus fetched artifacts needed for grading."""
-
-    trajectory: List[List[LettaMessageUnion]]
-    agent_id: Optional[str]
-    model_handle: Optional[str]
-    agent_usage: Optional[List[dict]]
-    agent_state: Optional[AgentState]
-    token_data: Optional[List[TurnTokenData]]
 
 
 def _model_id_for(model_handle: Optional[str]) -> str:
@@ -240,6 +227,16 @@ class Runner:
     def _should_cleanup_agent(self) -> bool:
         return self.suite.cleanup
 
+    async def _attach_requested_token_data(
+        self, result: SampleResult, return_token_data: bool
+    ) -> SampleResult:
+        """Attach host-fetched token data to a completed SampleResult when requested."""
+        if not return_token_data or not result.agent_id:
+            return result
+        return result.model_copy(
+            update={"token_data": await fetch_token_data(self.client, result.agent_id)}
+        )
+
     # ── setup ──
 
     async def _run_setup(self, model_handle: Optional[str] = None) -> None:
@@ -382,8 +379,7 @@ class Runner:
                         self.suite, sample, model_handle, t_sample_start
                     )
                     agent_id = result.agent_id
-                    if return_token_data and agent_id:
-                        result = result.model_copy(update={"token_data": await fetch_token_data(self.client, agent_id)})
+                    result = await self._attach_requested_token_data(result, return_token_data)
                     # Fire post-completion callbacks based on the final result —
                     # mid-sample events (grading_started, token streaming) are
                     # not emitted in v1 because the host only sees the final
